[frontend] fix layouts for narrow width + move dialogs to separate directory
