+++ frontend/src/dialog/AddDividendDialog.js
