+++ frontend/src/dialog/AddRecordDialog.js
