+++ frontend/src/dialog/AddTradeDialog.js
