+++ frontend/src/dialog/SellTradeDialog.js
