+++ frontend/src/views/Trades.js
@@ -2,8 +2,8 @@ import React, {useEffect, useState} from "react";
 import {Paper, Table, TableBody, TableCell, TableContainer, TableHead, TableRow} from "@mui/material";
 import {useData} from "../fetch";
 import Loader from "../components/Loader";
-import AddTradeDialog from "../components/AddTradeDialog";
-import SellTradeDialog from "../components/SellTradeDialog";
+import AddTradeDialog from "../dialog/AddTradeDialog";
+import SellTradeDialog from "../dialog/SellTradeDialog";
 
 
 function headerStyle(main, index){
